Rewrite MsgProto to be typecheckable
Stringly typed python be gone! This actually caught a fair few bugs in state.py with incorrectly named fields. This involves a little bit of hacky python and `__class__` assignments but I think it's acceptable as it improves the development experience so much.

Ideally all of the ProtobufMessage subclasses would be slotted but that doesn't work for 2 reasons:
1. betterproto doesn't support slotted classes and relies on usage of `__dict__` (this might change)
2. python doesn't support `__class__` assignments with slotted classes from a superclass (this probably won't change)

(This is also part of a larger set of commits so it's likely this won't actually work by itself)